Move access_roles enums to it's own module
